Move Premake code handling Examples into Examples/examples.lua
